move code that doesn't need timestamp cache out of lock
As far as I can see this method is only ever called via ILayout.Format
which in turn is only called via AppenderSkeleton.RenderLoggingEvent
which either calls it inside of a lock (the one-arg version in
AppenderSkeleton), uses a fresh instance of StringWriter (the SMTP
appenders) or is locked via the FileAppender's lock. So writer is
either not shared between threads at all or already protected by a
lock.