LOG4NET-461: refactored the AdoNetAppender
The AdoNetAppender suffered several issues. One was that a database command was
created and stored for the appenders lifetime. This caused trouble with some
database servers because those cleaned up cached prepared statements from time to
time and thus caused the appender to fail without the capability to recover. I
solved this problem by creating a database command whenever the buffer is sent,
well knowing that this costs a little more performance but is much more stable.

At this point I further took the chance to cleanup the source by replacing all
references to private member attributes with their public property mapping.
Further I added and fixed some of the documentation.

This overall cleanup is not yet complete, meaning that there is still a connection
that is being kept alive while it would be wiser to rely on proper connection
pooling and just open/close the database connection when sending the buffer. I
will raise this topic on the dev list.