Shutdown: Unsubscribe from AppDomain event handlers (#256)
* Shutdown: Unsubscribe from AppDomain event handlers

When log4net is used in an AssemblyLoadContext that should be unloaded, log4net blocks the unloading since it subscribed to event handlers from the default load context. This PR aims to cleanup the 2 subscriptions that most likely cause this behaviour.

Note that I did not fully test the solution yet. But this cleanup is a good measure in any case.